+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Warn the user when no code is loaded for a command that requires it
 * Change commands that run the developer tools to require the code to be loaded first
 * Inform the user when commands that spawn processes complete
+* Add experimental code lens support for test results
 * Add "Logtalk: Generate Documentation (workspace)" command
 * Add "Logtalk: Generate Diagrams (workspace)" command
 * Add "Logtalk: Scan Dead Code (workspace)" command